services/whisper-api/requirements.txt
openai-whisper==20231117
Flask==3.0.0